Extract create action
