+++ src/jlmkr/actions/list.py
@@ -8,8 +8,7 @@ import subprocess
 from collections import defaultdict
 from utils.console import NORMAL, UNDERLINE
 from utils.config_parser import parse_config_file
-from utils.jail_dataset import get_jail_config_path, get_jail_rootfs_path, parse_os_release
-from utils.parent_dataset import get_all_jail_names
+from utils.jail_dataset import get_all_jail_names, get_jail_config_path, get_jail_rootfs_path, parse_os_release
 
 
 def list_jails():
